refactor(866): move Stripe into src/Sections/Humans.Stripe (G5 lane 4b-2a)
Stands up the payments connector as its own section, per the 2026-08-14 end-state
design table (own section, not Base and not Store — the vendor-connector shape, with
TicketTailor and Holded as precedent).

Moved:
- Humans.Application/Interfaces/IStripeService.cs -> Contracts/IStripeService.cs
  (IStripeService + StripePaymentDetails + StoreCheckoutEventKind +
  StoreCheckoutWebhookEvent + StoreCheckoutSessionData, all public)
- Humans.Infrastructure/Services/{StripeService,StripeStartupSmokeService,
  StoreWebhookRegistrationService}.cs -> Services/, all internal sealed (HUM0034)
- Humans.Web/Extensions/Infrastructure/StripeInfrastructureExtensions.cs ->
  Section.cs verbatim; Shell's roll-call line is gone
- the three Stripe test files -> tests/Humans.Stripe.Tests

Contracts/ folder, no .Contracts leaf: Humans.Stripe references no section, so
Humans.Store -> Humans.Stripe and Humans.Tickets -> Humans.Stripe are acyclic and a
leaf would buy nothing.

No schema change: the section owns no tables, so no DbContext, no migrations, no
G4 gate.

Stripe.net leaves Humans.Infrastructure. That package was floating the whole
solution's transitive Newtonsoft.Json to 13.x over Octokit's 11.0.1; removing it
exposed the latent NU1903, so Directory.Packages.props takes a direct 13.0.4 pin on
the SSH.NET precedent already in that file.

Refs nobodies-collective/Humans#866
