+++ src/Sections/Humans.Store/Controllers/StoreStripeWebhookController.cs
@@ -1,5 +1,6 @@
 using Humans.Application.Interfaces;
 using Humans.Store.Services;
+using Humans.Stripe.Contracts;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
