+++ src/Sections/Humans.Stripe/Humans.Stripe.csproj
@@ -1,0 +1,50 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <InternalsVisibleTo Include="Humans.Stripe.Tests" />
+    <!-- The one assembly allowed to see many sections (design §5). -->
+    <InternalsVisibleTo Include="Humans.Integration.Tests" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <!--
+      The payments connector section (nobodies-collective/Humans#866 G5, lane 4b-2a;
+      placement decided at the 2026-08-14 end-state design session). Stripe.net stops
+      here: Contracts/IStripeService.cs is the §15i seam and no SDK type crosses it.
+
+      It owns no tables (no DbContext, no repository, no AddSectionDbContext — the Data
+      steps of G5-SECTION-TEMPLATE.md are skipped wholesale), no controllers and no
+      views, so this is plain Microsoft.NET.Sdk rather than Sdk.Razor (TicketTailor's
+      shape) and takes no Humans.Infrastructure reference (Scanner's shape, reached by a
+      connector that lived in Humans.Infrastructure/Services on the way in — Mailer's
+      case). Stripe fee values land on Tickets' TicketOrder and Store's Payment, both
+      written through those sections' own repositories.
+
+      Humans.Application is referenced for IApplicationService's sibling usings:
+      Humans.Application.Extensions (ILogger.TimeOperation) and
+      Humans.Application.Configuration.GitHubSettings, which the webhook registrar reads
+      to sweep endpoints for closed PRs.
+    -->
+    <ProjectReference Include="..\..\Humans.Interfaces\Humans.Interfaces.csproj" />
+    <ProjectReference Include="..\..\Humans.Domain\Humans.Domain.csproj" />
+    <ProjectReference Include="..\..\Humans.Application\Humans.Application.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <!--
+      No AspNetCore package here: those ship with no PackageVersion on purpose and would
+      fail the build with NU1010 under Central Package Management. IHostedService and
+      friends arrive through the framework reference below (design §15 step 1).
+    -->
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="NodaTime" />
+    <PackageReference Include="Octokit" />
+    <PackageReference Include="Stripe.net" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <!-- The section owns its documentation (design §7a); keep it visible in Solution Explorer. -->
+    <None Include="**\*.md" />
+  </ItemGroup>
+
+</Project>
