+++ src/Sections/Humans.Tickets/Humans.Tickets.csproj
@@ -44,6 +44,11 @@
     <ProjectReference Include="..\Humans.Campaigns.Contracts\Humans.Campaigns.Contracts.csproj" />
     <ProjectReference Include="..\Humans.Email.Contracts\Humans.Email.Contracts.csproj" />
     <ProjectReference Include="..\Humans.Gdpr.Contracts\Humans.Gdpr.Contracts.csproj" />
+    <!--
+      Stripe fee enrichment on TicketOrder, through Humans.Stripe.Contracts. Direct
+      section reference, acyclic — Humans.Stripe references no section.
+    -->
+    <ProjectReference Include="..\Humans.Stripe\Humans.Stripe.csproj" />
   </ItemGroup>
 
   <ItemGroup>
