+++ src/Sections/Humans.Tickets/Services/TicketSyncService.cs
@@ -14,6 +14,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using NodaTime;
 using Humans.Application.Interfaces.TicketVendor;
+using Humans.Stripe.Contracts;
 using Humans.Tickets.Data;
 using Humans.Tickets.Domain;
 using Humans.Users.Contracts;
