+++ tests/Humans.Store.Tests/Services/SummaryAggregateTests.cs
@@ -9,6 +9,7 @@ using Humans.Teams.Contracts;
 using Humans.Store.Data;
 using Humans.Store.Domain;
 using Humans.Store.Services;
+using Humans.Stripe.Contracts;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
 using Microsoft.Extensions.Logging.Abstractions;
