+++ tests/Humans.Tickets.Tests/Services/TicketSyncServiceNullOrderTests.cs
@@ -17,6 +17,7 @@ using Microsoft.Extensions.Options;
 using NodaTime;
 using NSubstitute;
 using Humans.Tickets.Domain;
+using Humans.Stripe.Contracts;
 using Humans.Users.Contracts;
 
 namespace Humans.Tickets.Tests.Services;
